@@ -15,6 +15,7 @@
     "@google-cloud/translate": "^8.3.0",
     "@types/pg": "^8.11.6",
     "axios": "^1.7.2",
+    "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
@@ -22,15 +23,18 @@
     "http-errors": "^2.0.0",
     "jsonwebtoken": "^9.0.2",
     "ms": "^2.1.3",
-    "pg": "^8.11.5",
-    "sequelize": "^6.37.3",
-    "typescript": "^5.4.5",
-    "uuid": "^9.0.1"
+    "pg": "^8.13.1",
+    "pg-hstore": "^2.3.4",
+    "sequelize": "^6.37.5",
+    "sequelize-typescript": "^2.1.6",
+    "uuid": "^9.0.1",
+    "zod": "^3.24.1"
   },
   "compilerOptions": {
     "outDir": "./dist"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/bcryptjs": "^2.4.6",
     "@types/cookie-parser": "^1.4.7",
     "@types/cors": "^2.8.17",
@@ -43,6 +47,7 @@
     "jest": "^29.7.0",
     "nodemon": "^3.1.1",
     "ts-jest": "^29.1.4",
-    "ts-node": "^10.9.2"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
   }
 }
