@@ -1,133 +1,56 @@
-// import express from "express"; // Assuming you're using Express for request handling
-// import jwt from "jsonwebtoken"; // For JSON Web Token generation
-// import createError from "http-errors";
-// import ms from "ms"; // For parsing time durations
-// import { saveAuthToken } from "../3_models/auth";
+import { Request, Response, NextFunction } from 'express';
+import jwt from 'jsonwebtoken';
+import { User } from '../3_models/user';
 
-// // Assuming you have a separate utility file for JWT generation
-// import { generateJWT } from "../utils/auth";
+// Extend Express Request type to include user
+declare global {
+  namespace Express {
+    interface Request {
+      user?: {
+        id: string;
+        iat?: number;
+        exp?: number;
+      };
+    }
+  }
+}
 
+const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key';
 
-// // Assuming you have environment variables set for these values
-// const ACCESS_TOKEN_LIFE: string = process.env.ACCESS_TOKEN_LIFE || "";
-// const REFRESH_TOKEN_LIFE: string = process.env.REFRESH_TOKEN_LIFE || "";
-// const ACCESS_TOKEN_SECRET: string = process.env.ACCESS_TOKEN_SECRET || "";
-// const REFRESH_TOKEN_SECRET: string = process.env.REFRESH_TOKEN_SECRET || "";
-// const NODE_ENV: string = process.env.NODE_ENV || "";
+export const generateAuthToken = (userId: string): string => {
+  return jwt.sign({ id: userId }, JWT_SECRET, { expiresIn: '24h' });
+};
 
-// interface Token {
-//   refreshToken: string;
-//   userId: string;
-//   expirationTime: number;
-// }
+export const isAuthenticated = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    // Get token from header
+    const authHeader = req.headers.authorization;
+    if (!authHeader?.startsWith('Bearer ')) {
+      return res.status(401).json({ message: 'No token provided' });
+    }
 
-// const dev = NODE_ENV === "development";
+    // Verify token
+    const token = authHeader.split(' ')[1];
+    const decoded = jwt.verify(token, JWT_SECRET) as { id: string };
 
-// import { Pool, PoolClient } from "pg";
-// import config from "../0_config/database";
+    // Get user from database
+    const user = await User.getUserById(decoded.id);
+    if (!user) {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
 
-// const pool = new Pool(config);
-
-// async function calculateExpirationTime(duration: string): Promise<number> {
-//   return Date.now() + ms(duration);
-// }
-
-// export async function generateAuthTokens(
-//   req: express.Request<any, any, { userId: string }>, // Use the extended Request type
-//   res: express.Response,
-//   next: express.NextFunction
-// ): Promise<void | express.Response<any, any>> {
-//   try {
-//     // Extract userId from the request (assuming it's stored somewhere)
-//     const userId = req.userId;
-
-//     // Find the user using userId
-//     const user = await User.getUserById(userId);
-
-//     if (!user) {
-//       throw createError.NotFound("User not found"); // Handle user not found error
-//     }
-
-//     const refreshToken = generateJWT(userId, REFRESH_TOKEN_SECRET, REFRESH_TOKEN_LIFE);
-//     const accessToken = generateJWT(userId, ACCESS_TOKEN_SECRET, ACCESS_TOKEN_LIFE);
-
-//     const expirationTime = await calculateExpirationTime(REFRESH_TOKEN_LIFE);
-
-//     const token: Token = {
-//       refreshToken,
-//       userId,
-//       expirationTime,
-//     };
-
-//     // Call saveAuthToken to store the token in the database
-//     await saveAuthToken(userId, refreshToken);
-
-//     res.cookie("refreshToken", refreshToken, {
-//       httpOnly: true,
-//       secure: !dev,
-//       signed: true,
-//       expires: new Date(expirationTime),
-//     });
-
-//     const expiresAt = new Date(Date.now() + ms(ACCESS_TOKEN_LIFE));
-
-//     return res.status(200).json({
-//       user,
-//       token: accessToken,
-//       expiresAt,
-//     });
-//   } catch (error) {
-//     return next(error);
-//   }
-// }
-
-// export const isAuthenticated = async (
-//     req: express.Request,  // Use the extended Request type
-//     res: express.Response,
-//     next: express.NextFunction
-//   ) => {
-//     try {
-//       const authToken = req.get('Authorization');
-//       const accessToken = authToken?.split('Bearer ')[1];
-//       if (!accessToken) {
-//         const error = createError.Unauthorized();
-//         throw error;
-//       }
-  
-//       const { signedCookies = {} } = req;
-//       const { refreshToken } = signedCookies;
-//       if (!refreshToken) {
-//         const error = createError.Unauthorized();
-//         throw error;
-//       }
-  
-//       let decodedToken;
-// try {
-//   decodedToken = jwt.verify(accessToken, ACCESS_TOKEN_SECRET);
-// } catch (err) {
-//   const error = createError.Unauthorized();
-//   return next(error);
-// }
-
-// const { userId } = decodedToken;
-
-// const user = await User.getUserById(userId);
-// if (!user) {
-//   const error = createError.Unauthorized();
-//   throw error;
-// }
-
-// // Assuming you want to set userId on the request object for further use
-// req.userId = user.id;
-  
-  
-//       return next();
-//     } catch (error) {
-//       return next(error);
-//     }
-//   };
-  
-//   export default {
-//     generateAuthTokens,
-//     isAuthenticated,
-//   };
+    // Add user to request object
+    req.user = decoded;
+    next();
+  } catch (error) {
+    if (error instanceof jwt.JsonWebTokenError) {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
+    console.error('Auth middleware error:', error);
+    res.status(500).json({ message: 'Internal server error' });
+  }
+};
